### 2019-01-17 #39 @lwjohnst86
@labarba Do all authors need to have an ORCID? She doesn't have one.
