### 2019-01-17 #39 @labarba
JOSE needs all authors to have an ORCID. Can you request that she get one?
