### 2019-01-26 #39 @labarba
👋  @lwjohnst86 — Did you talk with your co-author about getting an ORCID? We'll need that to process the paper. Let us know!
