### 2019-01-26 #39 @lwjohnst86
Yes, they finally responded just last night. I will updated now. I assume I just make that change to the repo and not have to make a new tag?
