### 2019-01-26 #39 @labarba
Just make the change in the paper metadata! Then you can run `@whedon generate pdf` yourself to feel the power of our bot.
