### 2019-04-20 #39 @labarba
👋  @K8hertweck — You signed up to review for JOSE a while back, marking R as language and biology as domain context ... could you have a look at this submission, and let me know if you'd be willing to contribute a review? Thank you!
