### 2019-04-20 #39 @labarba
Maybe @jsta could help us out? We're in need of reviewers for this submission to JOSE. Have a look and let me know if you're up for it ... Merci! 😎
