### 2019-04-23 #49 @whedon
[REVIEW]: A graduate student-led participatory live-coding quantitative methods course in R: Experiences on initiating, developing, and teaching
**Submitting author:** @lwjohnst86 (<a href="http://orcid.org/0000-0003-4169-2616">Luke W. Johnston</a>)
**Repository:** <a href="https://github.com/UofTCoders/rcourse" target ="_blank">https://github.com/UofTCoders/rcourse</a>
**Version:** v2.2.0
