### 2019-05-01 UofTCoders/rcourse#292 @jhollist
JOSE Review Comments
All,  very nice job on the course and the JOSE submission!

I am mostly finished with the review over at https://github.com/openjournals/jose-reviews/issues/49 but had a few specific comments to leave here.  Once these are addressed, I should be able to complete the review checklist.

As I mention on the JOSE issue, I though these course materials were excellent.  Will certainly be recommending them.

Any question about these comments, just let me know.

## Authorship
- Sara Mahallati is not listed on https://uoftcoders.github.io/rcourse/about.html but is listed as an author of the paper.  Should she be listed both places?

## Installation Instructions and Usage
- Create a separate section in README for installation.
- More explicit step by step instructions on how to create your own version of the course would be really useful.  Reason for rolling your own instructions is because some of what is suggested in the Bookdown docs, is taken care of in your .travis.yml and the gh-pages branch suggested in Bookdown isn't needed. Thus, as is, to get your own site up and running requires some prior experience with GitHub Pages and Travis.  I think having more specific instructions that are geared towards your repository would be beneficial
- In this section, might want to add "fork to you own repo" as a first step.
- Include some warning about how long the Travis-CI build will take.

## Community Guidelines
- The CONTRIBUTING.md is well done, but a little hidden under the **Lecture Content** section.  Create a separate section for this.

## References
- Add DOI to Wilson, 2006: https://doi.org/10.1109/MCSE.2006.122
- Add DOI to Van Burren and Groothuis-Oudshoorn 2011: http://dx.doi.org/10.18637/jss.v045.i03
- Add DOI to Hothorn et al 2008: https://doi.org/10.1002/bimj.200810425
- Make sure all the rest of the references that have DOIs have it listed.
