### 2019-05-01 #49 @jhollist
This is a very well done submission.  So well done in fact, that I feel odd in not providing more detailed suggestions on what to change.  I have gone back through the materials, the paper and the repository to see if I missed anything major, but I haven't.  The few things I did find and would like to see updated are listed in https://github.com/UofTCoders/rcourse/issues/292.  Biggest change that I suggest there is to provide more detailed installation instructions in the README.

Kudos to all of the authors on a job very well done.  Was a pleasure to review and I have been reading through these to learn some new tricks myself!
