### 2019-05-02 UofTCoders/rcourse#292 @lwjohnst86
Thanks for the comments and feedback @jhollist! We really appreciate them :) We'll work to address these changes!
