### 2019-05-14 e7f8dd4
Squashing rogue period in README.
