### 2019-05-14 #49 @MattForshaw
I would like to echo the earlier comments of @jhollist. This is a fantastic submission. The queries raised in https://github.com/UofTCoders/rcourse/issues/292 cover all improvements I can suggest. I also raise a very small PR with a punctuation fix. Once this issue closes I will be very happy to review the submission, and all being well, complete the checklist.
