### 2019-05-27 fc7a9a3
Move contributing section.
Resolves UofTCoders/paper-rcourse#73
