+++ README.md
@@ -50,8 +50,6 @@ The course's lesson material is broadly subdivided into three main topics:
 3. Reproducible science (Lectures 13-15)
     * Prepares students for project work period and introduces methods for reproducible science (GitHub, R Markdown).
 
-If you are interested in contributing to the course material, please refer to the guidelines in [CONTRIBUTING.md](https://github.com/UofTCoders/rcourse/blob/master/CONTRIBUTING.md). 
-
 If you are interested in using or modifying this content and repository for your 
 own course, you'll need to make a change to the Travis CI setup so you can host 
 the website on your own account. Check out the documentation on the [R Markdown
@@ -62,6 +60,11 @@ excerpt taken from the book:
 1. Encrypt it in the environment variable `GITHUB_PAT` via command line `travis encrypt` and store it in `.travis.yml`, 
 e.g `travis encrypt GITHUB_PAT=TOKEN`. If you do not know how to install or use the Travis command-line tool, simply save this environment variable via https://travis-ci.org/user/repo/settings where `user` is your GitHub ID, and `repo` is the name of the repository.
 1. You can clone this `gh-pages` branch on Travis using your GitHub token, add the HTML output files from R Markdown (do not forget to add figures and CSS style files as well), and push to the remote repository.
+
+    
+## Contributing
+
+If you are interested in contributing to the course material, please refer to the guidelines in [CONTRIBUTING.md](https://github.com/UofTCoders/rcourse/blob/master/CONTRIBUTING.md). 
 
 ## Related Publications
 
