### 2019-05-27 ab25741
Add step by step instruction for setting up own forked repo
Resolves UofTCoders/paper-rcourse#74
