### 2019-05-30 7ff0dda
update with DOIs
