+++ paper.bib
@@ -84,6 +84,7 @@
 	publisher = {Cambridge University Press},
 	editor = {Sawyer, R. Keith},
 	year = {2006},
+	doi = {10.1192/bjp.bp.106.029678},
 	note = {OCLC: ocm62728545},
 	keywords = {Cognitive learning, Learning, Learning, Psychology of, Social aspects}
 }
@@ -94,6 +95,7 @@
     journal = {Computing in Science \& Engineering},
     month =   {November--December},
     year =    {2006},
+    doi = {10.1109/MCSE.2006.122},
     note =    {Summarizes the what and why of Version 3 of the course.}
 }
 
@@ -180,6 +182,7 @@
   volume = {45},
   number = {3},
   pages = {1-67},
+  doi = {10.18637/jss.v045.i03},
   url = {https://www.jstatsoft.org/v45/i03/},
 }
 
@@ -191,6 +194,7 @@
   volume = {50},
   number = {3},
   pages = {346--363},
+  doi = {10.1002/bimj.200810425}
 }
 
 @Manual{psych,
@@ -222,6 +226,7 @@
   number = {12},
   pages = {1--20},
   url = {https://www.jstatsoft.org/v21/i12/},
+  doi = {10.18637/jss.v021.i12},
 }
 
 @Article{plyr,
@@ -233,6 +238,7 @@
   number = {1},
   pages = {1--29},
   url = {https://www.jstatsoft.org/v40/i01/},
+  doi = {10.18637/jss.v040.i01}
 }
 
 @Manual{MuMIn,
@@ -274,5 +280,5 @@
   year = {2015},
   note = {R package version 0.1.0},
   url = {https://github.com/gotellilab/EcoSimR},
-  doi = {10.5281/zenodo.16522},
+  doi = {10.5281/zenodo.16522}
 }
