+++ paper.bib
@@ -41,7 +41,8 @@
 	language = {en},
 	publisher = {Amazon Distribution GmbH},
 	author = {Wilson, Greg},
-	year = {2018}
+	year = {2018},
+	url = {http://teachtogether.tech/}
 }
 
 @article{strobel_when_2009,
