### 2019-05-31 UofTCoders/rcourse#292 @lwjohnst86
@jhollist Sorry for the delay. We've updated the repo as per your requests!
