### 2019-05-31 #49 @lwjohnst86
@jhollist and @MattForshaw we've updated the repo as per https://github.com/UofTCoders/rcourse/issues/292. Thanks for the comments and kind words!
