### 2019-06-04 UofTCoders/rcourse#292 @jhollist
@lwjohnst86 Looks good!  Heading over the finish up my checklist now.

Congrats, this is a really nice looking course!
