### 2019-06-04 #49 @jhollist
I agree.  I recommend acceptance.  It is a nice paper and a great looking
class!

On Tue, Jun 4, 2019, 3:00 PM Matthew Forshaw <notifications@github.com>
wrote:

> Hi @labarba <https://github.com/labarba>. I'm very happy to recommend
> this paper for acceptance.
>
