### 2019-06-05 #49 @juanklopper
@labarba I see all the boxes are ticked.  I see @lwjohnst86 's archive in Zenodo with the DOI.  (Is this correct @lwjohnst86 DOI 10.5281/zenodo.2540708 ?)
