### 2019-06-05 #49 @labarba
It looks like v2.1.0 is from Jan 15 — we are going to need a new tagged release with the reviewed version, and the Zenodo archive to be updated to the latest release.
