### 2019-06-05 #49 @labarba
@lwjohnst86 — In the paper, you have the sentence "We believe the included lesson topics, open accessibility, and modularity of our course makes it an ideal resource for instructors and an excellent contribution to JOSE." I agree with it, of course, as the reverse and editor have accepted your submission. But it seems awkward to have such a statement _in the paper_. I suggest deleting it
