### 2019-06-05 #49 @labarba
You have this reference:

Johnston, L. W., Ostblom, J., Bonsma-Fisher, M., Santangelo, J. S., Hasan, A., Andrade, E. S. de, Coome, L., et al. (2018). UofTCoders/rcourse: EEB313 quantitative methods in R for biology. UofT Coders. doi:10.5281/zenodo.2335179

... if this is the Zenodo deposit of this very learning module, we ask that you don't include this as a citation, because the archive will be highlighted in the margin decorators on the first page of the paper (and the DOIs will be linked in the metadata of the paper).
