### 2019-06-05 #49 @labarba
I've read through the paper, and don't have further comments (except I'd add a comma after "e.g.")—we're ready to publish. At this time, we'll need the new version and the Zenodo deposit. Do report here the version number and the Zenodo DOI.
