+++ paper.bib
@@ -127,16 +127,6 @@
   doi = {10.5281/zenodo.569338},
 }
 
-@Manual{rcourse,
-  title = {{UofTCoders}/rcourse: {EEB313} Quantitative Methods in {R} for Biology},
-  author = {Luke W. Johnston and Joel Ostblom and Madeleine Bonsma-Fisher and James S. Santangelo and Ahmed Hasan and Elliott Sales de Andrade and Lindsay Coome and Lina Tran},
-  organization = {UofT Coders},
-  year = {2018},
-  note = {version 2.0.0},
-  doi = {10.5281/zenodo.2335179},
-  url = {https://doi.org/10.5281/zenodo.2335179},
-}
-
 @Manual{dplyr,
   title = {dplyr: A grammar of data manipulation},
   author = {Hadley Wickham and Romain François and Lionel Henry and Kirill Müller},
