### 2019-06-06 c5a0941
update from paper-rcourse repo
