### 2019-06-06 #49 @lwjohnst86
I've removed the reference as requested and updated the version.

Version: v2.2.0
Zenodo DOI: 10.5281/zenodo.3240508

Thanks!
