### 2019-06-06 #49 @labarba
Congratulations, @lwjohnst86, your paper is published in JOSE! 🚀

Tremendous heartfelt thanks to our editor: @juanklopper, and the reviewers: @jhollist, @MattForshaw 🙏
